Improved version of memset
Similar 20x improvement as the change to memmove

Also added profile scripts and changed error message in fileop

+++ sys/include/libc.h
@@ -23,7 +23,7 @@ extern void (*_abort)(void);
  * mem routines
  */
 extern	void*	memccpy(void*, const void*, int, uint32_t);
-extern	void*	memset(void*, int, uint32_t);
+extern	void*	memset(void*, int, size_t);
 extern	int	memcmp(const void*, const void*, uint32_t);
 extern	void*	memcpy(void*, const void*, size_t);
 extern	void*	memmove(void*, const void*, size_t);
